feat(devtools-remote): relocate Ktor WS transport + RemoteConfig + RemoteOutput
