+++ redux-kotlin-devtools-remote/src/commonMain/kotlin/org/reduxkotlin/devtools/WebSocketEngine.kt
@@ -1,0 +1,6 @@
+package org.reduxkotlin.devtools.remote
+
+import io.ktor.client.HttpClient
+
+/** Builds the platform [HttpClient] with the WebSockets plugin installed. */
+internal expect fun createDevToolsHttpClient(): HttpClient
