+++ redux-kotlin-devtools-remote/src/wasmJsMain/kotlin/org/reduxkotlin/devtools/WebSocketEngine.wasmJs.kt
@@ -1,0 +1,9 @@
+package org.reduxkotlin.devtools.remote
+
+import io.ktor.client.HttpClient
+import io.ktor.client.engine.js.Js
+import io.ktor.client.plugins.websocket.WebSockets
+
+internal actual fun createDevToolsHttpClient(): HttpClient = HttpClient(Js) {
+    install(WebSockets)
+}
