--- redux-kotlin-devtools/src/commonMain/kotlin/org/reduxkotlin/devtools/WebSocketEngine.kt
@@ -1,6 +1,0 @@
-package org.reduxkotlin.devtools
-
-import io.ktor.client.HttpClient
-
-/** Builds the platform [HttpClient] with the WebSockets plugin installed. */
-internal expect fun createDevToolsHttpClient(): HttpClient
